requirements.txt
hopsworks==4.2.*
pandas>=1.3.0
scikit-learn>=1.0.0
xgboost>=1.5.0
joblib>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
python-dotenv>=0.19.0